requirements.txt
-i https://pypi.org/simple
asn1crypto==0.24.0
attrs==19.1.0
automat==0.7.0
cffi==1.12.3
constantly==15.1.0
cryptography==2.7
cssselect==1.1.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
lxml==4.4.1 ; python_version != '3.4'
parsel==1.5.2
pyasn1-modules==0.2.6
pyasn1==0.4.7
pycparser==2.19
pydispatcher==2.0.5
pyhamcrest==1.9.0
pymysql==0.9.3
pyopenssl==19.0.0
python-dotenv==0.10.3
pywin32==225 ; sys_platform == 'win32'
queuelib==1.5.0
scrapy-magicfields==1.1.0
scrapy-mysql-pipeline==2019.7.19
scrapy==1.7.3
service-identity==18.1.0
six==1.12.0
twisted==19.7.0 ; python_version != '3.4'
w3lib==1.21.0
zope.interface==4.6.0
